Add support for multi-word person names
Names of payers and recipients can now be a string, not just a single
word. In the current state, this breaks the import of transactions from
a file, if any name in the file is longer than a single word.

Also fixes a bug where giving the same name in both name fields while
adding a transaction printed two (pridivide) prompts.